Add initial database schema, implement BigDecimal for monetary values, and enhance error handling
- Create initial SQL migration for database schema including tables for product, customer, merchant, courier, orders, and order_item.
- Update model classes to use BigDecimal for price-related fields to ensure precision.
- Implement GlobalExceptionHandler for centralized exception management and improved API error responses.
- Add Swagger/OpenAPI configuration for API documentation and testing.
- Introduce Flyway for database migrations and versioning.
- Document project architecture and concurrency handling strategies.

+++ orders-api/src/main/java/com/example/orders_api/service/ProductSearchService.java
@@ -1,8 +1,9 @@
 package com.example.orders_api.service;
 
 import com.example.orders_api.model.Product;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductSearchService {
-    List<Product> search(String name, Double minPrice, Double maxPrice);
+    List<Product> search(String name, BigDecimal minPrice, BigDecimal maxPrice);
 }
